// src/openapi/generated/models/voice-parse-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Esquema para respuesta de parseo de voz exitoso
 */
export interface VoiceParseResponse {
  confirmation_message?: (string | null);
  error?: (string | null);
  resident_id?: (string | null);
  resident_name?: (string | null);
  status?: (string | null);
  success: boolean;
  task_id?: (string | null);
  task_name?: (string | null);
}